requirements.txt
psycopg2-binary
python-dotenv
pandas
requests
beautifulsoup4
markdown
sqlalchemy
fastapi
uvicorn
pinecone
sentence-transformers
openai
scikit-learn
numpy<2.0
